server/requirements.txt
# Core backend
fastapi
fastapi-cors
uvicorn[standard]
uvicorn
python-docx
PyPDF2
openai
python-dotenv
aiofiles  
python-multipart
requests
httpx
cors
SQLAlchemy
psycopg2-binary
alembic
pydantic
pydantic_settings
pyjwt
passlib[bcrypt]
python-multipart
fastapi[standard]